requirements.txt
python-dotenv==1.0.1
langchain_community==0.3.4
langchain_openai==0.2.5
langgraph==0.2.45
google-search-results==2.4.2
streamlit==1.42.2
questionary==2.0.1